requirements.txt
# requirements.txt

pandas==2.1.4
numpy==1.23.5
wandb==0.16.1
tqdm==4.66.1
pytorch_lightning==2.1.2
transformers[torch]==4.35.2
rouge==1.0.1
jupyter==1.0.0
jupyterlab==4.0.9
hydra-core==1.3.2
rich>=10.2.2
evaluate
rouge_score
nltk
absl-pyhydra-core==1.3.2
scikit-learn
konlpy